// src/generated/api-types.ts
export interface components {
    schemas: {
// ...
        ClosedDateResponseDto: {
            /** Format: date */
            date: string;
            reason?: string;
        };
// ...
        CreateBlockDto: {
            date: string;
            start: string;
            end: string;
            reason: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}